pyproject.toml
[project]
name = "line-translation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "firebase-admin>=7.3.0",
    "flask>=3.1.3",
    "google-genai>=1.67.0",
    "gunicorn>=25.1.0",
    "line-bot-sdk>=3.22.0",
]
